mcp_servers/fred_connector/pyproject.toml
[project]
name = "fred-connector"
version = "0.1.0"
description = "Read-only FRED (St. Louis Fed) connector exposed as an MCP server"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
]

[project.scripts]
fred-connector = "fred_connector.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fred_connector"]
